pi_daemon/requirements.txt
RPi.GPIO
python-daemon
flask
flask_restful
